Update And Fix Interaction APIs
-Fix getall api.
-Add addition field of get response.
-Add checking id when call create/update apis.

+++ OnComics.BE/OnComics.Application/Services/Implements/InteractionService.cs
@@ -1,9 +1,11 @@
 ﻿using MapsterMapper;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.IdentityModel.Tokens;
 using OnComics.Application.Enums.Interaction;
 using OnComics.Application.Models.Request.Interaction;
 using OnComics.Application.Models.Response.Common;
 using OnComics.Application.Models.Response.Interaction;
+using OnComics.Application.Models.Response.InteractionType;
 using OnComics.Application.Services.Interfaces;
 using OnComics.Infrastructure.Entities;
 using OnComics.Infrastructure.Repositories.Interfaces;
@@ -15,13 +17,19 @@ namespace OnComics.Application.Services.Implements
     public class InteractionService : IInteractionService
     {
         private readonly IInteractionRepository _interactionRepository;
+        private readonly IInteractionTypeRepository _interactionTypeRepository;
+        private readonly ICommentRepository _commentRepository;
         private readonly IMapper _mapper;
 
         public InteractionService(
             IInteractionRepository interactionRepository,
+            IInteractionTypeRepository interactionTypeRepository,
+            ICommentRepository commentRepository,
             IMapper mapper)
         {
             _interactionRepository = interactionRepository;
+            _interactionTypeRepository = interactionTypeRepository;
+            _commentRepository = commentRepository;
             _mapper = mapper;
         }
 
@@ -37,18 +45,34 @@ namespace OnComics.Application.Services.Implements
 
                 bool isDecending = getInteractionReq.IsDescending;
 
-                Guid? searchId = getInteractionReq.AccountId;
+                Guid? searchId = getInteractionReq.Id;
+
+                bool? isAccId = getInteractionReq.IdType switch
+                {
+                    InteractionIdType.ACCOUNT => true,
+                    InteractionIdType.COMMENT => false,
+                    _ => null
+                };
 
                 Expression<Func<Interaction, bool>>? search = null;
 
                 int totalData = 0;
 
-                if (searchId.HasValue)
+                if (searchId.HasValue && isAccId == true)
                 {
                     search = i =>
                         (string.IsNullOrEmpty(searchKey) ||
-                        EF.Functions.Like(i.Comment.Account.Fullname, $"%{search}%") &&
-                        i.AccountId == searchId);
+                        EF.Functions.Like(i.Account.Fullname, $"%{search}%")) &&
+                        i.AccountId == searchId;
+
+                    totalData = await _interactionRepository.CountInteractionAsync(searchId);
+                }
+                else if (searchId.HasValue && isAccId == false)
+                {
+                    search = i =>
+                        (string.IsNullOrEmpty(searchKey) ||
+                        EF.Functions.Like(i.Comment.Account.Fullname, $"%{search}%")) &&
+                        i.CommentId == searchId;
 
                     totalData = await _interactionRepository.CountInteractionAsync(searchId);
                 }
@@ -62,21 +86,22 @@ namespace OnComics.Application.Services.Implements
                     totalData = await _interactionRepository.CountInteractionAsync();
                 }
 
-                Func<IQueryable<Interaction>, IOrderedQueryable<Interaction>>? order = r => getInteractionReq.SortBy switch
-                {
-                    InteractionSortOption.ACCOUNT => isDecending
-                        ? r.OrderByDescending(r => r.Account.Fullname)
-                        : r.OrderBy(r => r.Account.Fullname),
-                    InteractionSortOption.TIME => isDecending
-                        ? r.OrderByDescending(r => r.ReactTime)
-                        : r.OrderBy(r => r.ReactTime),
-                    _ => r.OrderBy(r => r.Id)
-                };
+                Func<IQueryable<Interaction>, IOrderedQueryable<Interaction>>? order = r =>
+                    getInteractionReq.SortBy switch
+                    {
+                        InteractionSortOption.ACCOUNT => isDecending
+                            ? r.OrderByDescending(r => r.Account.Fullname)
+                            : r.OrderBy(r => r.Account.Fullname),
+                        InteractionSortOption.TIME => isDecending
+                            ? r.OrderByDescending(r => r.ReactTime)
+                            : r.OrderBy(r => r.ReactTime),
+                        _ => r.OrderBy(r => r.Id)
+                    };
 
-                var (interactions, accounts, comments) = await _interactionRepository
+                var (interactions, accounts, comments, types) = await _interactionRepository
                     .GetInteractionsAsync(search, order, pageNum, pageIndex);
 
-                if (interactions == null)
+                if (interactions == null || interactions.IsNullOrEmpty())
                     return new ObjectResponse<IEnumerable<InteractionRes>?>(
                         (int)HttpStatusCode.NotFound,
                         "Interaction Data Empty!");
@@ -84,11 +109,11 @@ namespace OnComics.Application.Services.Implements
                 var data = interactions.Select(i => new InteractionRes
                 {
                     Id = i.Id,
-                    AccountId = i.AccountId,
-                    Fullname = accounts[i.AccountId],
-                    CommentId = i.CommentId,
-                    CommentAuthor = comments[i.CommentId],
-                    TypeId = i.TypeId,
+                    AccountId = accounts![i.Id].Item1,
+                    Fullname = accounts![i.Id].Item2,
+                    CommentId = comments![i.Id].Item1,
+                    CommentAuthor = comments![i.Id].Item2,
+                    Type = _mapper.Map<InteractionTypeRes>(types![i.Id]),
                     ReactTime = i.ReactTime
                 });
 
@@ -115,7 +140,7 @@ namespace OnComics.Application.Services.Implements
         {
             try
             {
-                var (interaction, fullname, author) = await _interactionRepository
+                var (interaction, fullname, author, type) = await _interactionRepository
                     .GetInteractionById(id);
 
                 if (interaction == null)
@@ -126,6 +151,7 @@ namespace OnComics.Application.Services.Implements
                 var data = _mapper.Map<InteractionRes>(interaction);
                 data.Fullname = fullname;
                 data.CommentAuthor = author;
+                data.Type = _mapper.Map<InteractionTypeRes>(type!);
 
                 return new ObjectResponse<InteractionRes?>(
                     (int)HttpStatusCode.OK,
@@ -144,19 +170,42 @@ namespace OnComics.Application.Services.Implements
         //Create Interaction
         public async Task<ObjectResponse<Interaction>> CreateInteractionAsync(Guid accId, CreateInteractionReq createInteractionReq)
         {
+            var newItr = new Interaction();
+
             try
             {
+                var comment = await _commentRepository
+                    .GetByIdAsync(createInteractionReq.CommentId, true);
+
+                if (comment == null)
+                    return new ObjectResponse<Interaction>(
+                        (int)HttpStatusCode.NotFound,
+                        "Comment Not Found!");
+
+                var type = await _interactionTypeRepository
+                    .GetByIdAsync(createInteractionReq.TypeId, false);
+
+                if (type == null)
+                    return new ObjectResponse<Interaction>(
+                        (int)HttpStatusCode.NotFound,
+                        "Interaction Type Not Found!");
+
                 var isExisted = await _interactionRepository
                     .CheckInteractionExistedAsync(accId, createInteractionReq.CommentId);
 
                 if (isExisted)
                     return new ObjectResponse<Interaction>(
-                        (int)HttpStatusCode.NotFound,
+                        (int)HttpStatusCode.BadRequest,
                         "Interaction Is Existed!");
 
-                var newItr = _mapper.Map<Interaction>(createInteractionReq);
+                newItr = _mapper.Map<Interaction>(createInteractionReq);
+                newItr.AccountId = accId;
+
+                comment.InteractionNum = comment.InteractionNum + 1;
 
                 await _interactionRepository.InsertAsync(newItr);
+
+                await _commentRepository.UpdateAsync(comment);
 
                 return new ObjectResponse<Interaction>(
                     (int)HttpStatusCode.Created,
@@ -165,6 +214,11 @@ namespace OnComics.Application.Services.Implements
             }
             catch (Exception ex)
             {
+                var interaction = await _interactionRepository.GetByIdAsync(newItr.Id, true);
+
+                if (interaction != null)
+                    await _interactionRepository.DeleteAsync(interaction);
+
                 return new ObjectResponse<Interaction>(
                     (int)HttpStatusCode.InternalServerError,
                     ex.GetType().FullName!,
@@ -177,6 +231,14 @@ namespace OnComics.Application.Services.Implements
         {
             try
             {
+                var type = await _interactionTypeRepository
+                    .GetByIdAsync(updateInteractionReq.TypeId, false);
+
+                if (type == null)
+                    return new VoidResponse(
+                        (int)HttpStatusCode.NotFound,
+                        "Interaction Type Not Found!");
+
                 var oldItr = await _interactionRepository.GetByIdAsync(id, true);
 
                 if (oldItr == null)
@@ -204,14 +266,25 @@ namespace OnComics.Application.Services.Implements
         //Delete Interaction
         public async Task<VoidResponse> DeleteInteractionAsync(Guid id)
         {
+            var interaction = new Interaction();
+
             try
             {
-                var interaction = await _interactionRepository.GetByIdAsync(id, true);
+                interaction = await _interactionRepository.GetByIdAsync(id, true);
 
                 if (interaction == null)
                     return new VoidResponse(
                         (int)HttpStatusCode.NotFound,
                         "Interaction Not Found!");
+
+                var comment = await _commentRepository.GetByIdAsync(interaction.CommentId, true);
+
+                if (comment == null)
+                    return new VoidResponse(
+                        (int)HttpStatusCode.NotFound,
+                        "Comment Not Found!");
+
+                comment.InteractionNum = comment.InteractionNum - 1;
 
                 await _interactionRepository.DeleteAsync(interaction);
 
@@ -221,6 +294,11 @@ namespace OnComics.Application.Services.Implements
             }
             catch (Exception ex)
             {
+                var data = await _interactionRepository.GetByIdAsync(id, false);
+
+                if (data == null && interaction != null)
+                    await _interactionRepository.InsertAsync(interaction);
+
                 return new VoidResponse(
                     (int)HttpStatusCode.InternalServerError,
                     ex.GetType().FullName!,
